Change format from controller to middleware
